Corrections from printer.
1. We used the wrong style file. Fixed.

2. Page size was wrong in submitted PDF. This could have to do with Greg's
LaTeX environment... I did a visual check against an ACM-sanctioned example and
we're now good. (That is, we're no longer "high" on the page.) I am running a
recent version of LaTeX on my Mac, which might be the difference? Either way,
"Preview" says we're 8.5 x 11, and visual says we're good.

3. Put in a page break so that section 3 started at top of second page. (This
was possible because the new style file bought us some space.)

That seems to be it.